requirements.txt
Django==4.2.3
Pillow==9.5.0
PyWavelets==1.4.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
contourpy==1.1.0
cycler==0.11.0
fonttools==4.41.0
igraph==0.10.4
image==1.5.33
imageio==2.31.1
importlib-resources==6.0.0
joblib==1.3.1
kiwisolver==1.4.4
lazy_loader==0.3
matplotlib==3.7.0
networkx==3.1
numpy==1.23.5
opencv-python==4.7.0.68
packaging==23.1
pandas==2.0.0
pkg_resources==0.0.0
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
shapely==2.0.1
six==1.16.0
sqlparse==0.4.4
texttable==1.6.7
threadpoolctl==3.2.0
tifffile==2023.7.10
tqdm==4.65.1
typing_extensions==4.7.1
tzdata==2023.3
zipp==3.16.2
